refactor: add src directory
