+++ src/__test__/index.test.tsx
@@ -1,11 +1,11 @@
 // __tests__/index.test.jsx
 
-import Header from '@/components/Common/Header';
-import Home from '@/pages/index';
 import '@testing-library/jest-dom';
 import { login } from 'mocks/handlers';
 import { server } from 'mocks/server';
 import { act } from 'react-dom/test-utils';
+import Header from 'src/components/Common/Header';
+import Home from 'src/pages/index';
 import { fireEvent, render, screen } from './test-utils';
 
 describe('Header and navigations', () => {
