+++ src/components/Projects/SmallCard.tsx
